Moved split event array tests into their own Tests file
