+++ Tests/InkableTests/PolylineTests.swift
@@ -127,82 +127,4 @@ class PolylineTests: XCTestCase {
         XCTAssertEqual(polylineOutput.deltas[0], .updatedPolyline(index: 0, updatedIndexes: IndexSet([0, 1, 2])))
         XCTAssertEqual(polylineOutput.deltas[1], .completedPolyline(index: 0))
     }
-
-    func testStreamsMatch() throws {
-        guard
-            let jsonFile = Bundle.module.url(forResource: "pencil-error", withExtension: "json")
-        else {
-            XCTFail("Could not load json")
-            return
-        }
-
-        let data = try Data(contentsOf: jsonFile)
-        let events = try JSONDecoder().decode([TouchEvent].self, from: data)
-        let touchStream = TouchPathStream()
-        let polylineStream = PolylineStream()
-        touchStream.addConsumer(polylineStream)
-        touchStream.produce(with: events)
-
-        for split in 1..<events.count {
-            let altStream = TouchPathStream()
-            let altPolylineStream = PolylineStream()
-            altStream.addConsumer(altPolylineStream)
-            altStream.produce(with: Array(events[0 ..< split]))
-            altStream.produce(with: Array(events[split ..< events.count]))
-
-            XCTAssertEqual(polylineStream.lines, altPolylineStream.lines)
-        }
-    }
-
-    func testStreamsMatch2() throws {
-        guard
-            let jsonFile = Bundle.module.url(forResource: "unknown-polyline-index", withExtension: "json")
-        else {
-            XCTFail("Could not load json")
-            return
-        }
-
-        let data = try Data(contentsOf: jsonFile)
-        let events = try JSONDecoder().decode([TouchEvent].self, from: data)
-        let touchStream = TouchPathStream()
-        let polylineStream = PolylineStream()
-        touchStream.addConsumer(polylineStream)
-        touchStream.produce(with: events)
-
-        for split in 1..<events.count {
-            let altStream = TouchPathStream()
-            let altPolylineStream = PolylineStream()
-            altStream.addConsumer(altPolylineStream)
-            altStream.produce(with: Array(events[0 ..< split]))
-            altStream.produce(with: Array(events[split ..< events.count]))
-
-            XCTAssertEqual(polylineStream.lines, altPolylineStream.lines)
-        }
-    }
-
-    func testStreamsMatch3() throws {
-        guard
-            let jsonFile = Bundle.module.url(forResource: "pencil-antigrain", withExtension: "json")
-        else {
-            XCTFail("Could not load json")
-            return
-        }
-
-        let data = try Data(contentsOf: jsonFile)
-        let events = try JSONDecoder().decode([TouchEvent].self, from: data)
-        let touchStream = TouchPathStream()
-        let polylineStream = PolylineStream()
-        touchStream.addConsumer(polylineStream)
-        touchStream.produce(with: events)
-
-        for split in 1..<events.count {
-            let altStream = TouchPathStream()
-            let altPolylineStream = PolylineStream()
-            altStream.addConsumer(altPolylineStream)
-            altStream.produce(with: Array(events[0 ..< split]))
-            altStream.produce(with: Array(events[split ..< events.count]))
-
-            XCTAssertEqual(polylineStream.lines, altPolylineStream.lines)
-        }
-    }
 }
