+++ Tests/InkableTests/TouchPathTests.swift
@@ -80,28 +80,6 @@ class TouchPathTests: XCTestCase {
         }
     }
 
-    func testStreamsMatch2() throws {
-        guard
-            let jsonFile = Bundle.module.url(forResource: "events", withExtension: "json")
-        else {
-            XCTFail("Could not load json")
-            return
-        }
-
-        let data = try Data(contentsOf: jsonFile)
-        let events = try JSONDecoder().decode([TouchEvent].self, from: data)
-        let touchStream = TouchPathStream()
-        touchStream.produce(with: events)
-
-        for split in 1..<events.count {
-            let altStream = TouchPathStream()
-            altStream.produce(with: Array(events[0 ..< split]))
-            altStream.produce(with: Array(events[split ..< events.count]))
-
-            XCTAssertEqual(touchStream.paths, altStream.paths)
-        }
-    }
-
     func testMeasureTouchEvents() throws {
         guard
             let jsonFile = Bundle.module.url(forResource: "events", withExtension: "json")
